+++ CHANGELOG.md
@@ -9,7 +9,8 @@ NOTE: all lowercase and no ending punctuation marks
 
 ## Next release v1.0
 
-- [RY] - [#61](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/61) app(feature): status-for-specific-aud
+- [RY] - [#64](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/64) app(hotfix): bubble display fix
+- [RY] - [#62](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/62) app(feature): status-for-specific-aud
 - [RY] - [#60](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/60) app(feature): bubble-edit-backend
 - [RY] - [#59](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/59) app(feature): bubble backend
 - [CY] - [#40](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/40) app(feature): chat info backend